refactor: add common folder
