+++ src/App.tsx
@@ -2,11 +2,12 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { hot } from 'react-hot-loader/root';
 import React, { Suspense, lazy } from 'react';
 import styled from 'styled-components';
-import GlobalStyle from '@components/GlobalStyle';
-import IconLink from '@components/iconLink';
 import HomeIcon from '@mui/icons-material/Home';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
+import GlobalStyle from '@components/GlobalStyle';
 import Loading from '@routes/Loading';
+import IconLink from '@common/iconLink';
+import Sidebar from '@common/sidebar';
 
 // TODO: type definition 다른 폴더로 빼기
 declare global {
@@ -25,13 +26,13 @@ const App = (): JSX.Element => (
 	<Router>
 		<GlobalStyle />
 		<Wrapper>
-			<SideBar>
+			<Sidebar>
 				<IconLink to="/" icon={<HomeIcon fontSize="large" />} />
 				<IconLink
 					to="/login"
 					icon={<AccountCircleIcon fontSize="large" />}
 				/>
-			</SideBar>
+			</Sidebar>
 			<Switch>
 				<Suspense fallback={<Loading />}>
 					<Route path="/" exact component={Home} />
@@ -44,18 +45,6 @@ const App = (): JSX.Element => (
 		</Wrapper>
 	</Router>
 );
-
-const SideBar = styled('div')({
-	zIndex: 10,
-	display: 'flex',
-	flexDirection: 'column',
-	width: '70px',
-	height: '100%',
-	padding: '40px 0px',
-	boxShadow: 'rgba(17, 17, 26, 0.1) 3px 3px 4px 1px',
-
-	backgroundColor: '#F4F0EF',
-});
 
 const Wrapper = styled('div')({
 	display: 'flex',
