+++ tsconfig.json
@@ -49,6 +49,7 @@
     "baseUrl": "./",                       /* Base directory to resolve non-absolute module names. */
     "paths": {
       "@mui/styled-engine": ["./node_modules/@mui/styled-engine-sc"],
+      "@common/*": ["src/components/common/*"],
       "@components/*": ["src/components/*"],
       "@library/*": ["src/library/*"],
       "@routes/*": ["src/routes/*"],
